renaming config file to env.js
